requirements.txt
requests
PyPDF2
langchain_google_genai
brotli
brotlicffi
python-dotenv
